Made start state singular.
Changed start state in automata to singular value (string) because it
doesn't make sense for it to be a list, there will always be only one
start state. Corrected corresponding method and references to that method.

Also included in this commit are improvements to the default values for
creating Automatas. Found a pitfall where default values are actually only
evaluated once, so if you modified them somehow, it carried across to all
instantiations of the object (!!!).