Cleared up naming issues by renaming files.
As we need to produce a scanner object that will tokenize a string, scanner.py needed to be renamed to reflect that it does implementation-specific lexing. It is now description_reader.py as its job is specifically to lex descriptions for use by the project. lexical_desc.py turned out to be the object that most closely resembles the scanner object that we need to actually produce, so it was named scanner.py.